Win SDK FlipView sample now working in bluesky
The FlipView sample is working for all but 1 of the test scenarios
(which requires VirtualizedDataSources which hasn't been implemented
yet).  There are various minor bugs which are noted in the code/UX, but
overall it's the Win8 FlipView sample running on the web.  The main bug
to be tackled is a minor flash of partial content when changing
scenarios; it's because the Page rendering code needs to wait until all
subpages are rendered before adding the Page to the DOM.

This commit also adds:
* bugfixes in Animation and Pages
* "bluesky-MiscJSPolyfills.js" - contains various wrappers to make
IE10-only features/properties (e.g. innerText, onpropertychange) work on
other browsers.  These functions althrough throw one warning when an
IE10-only feature is hit, to notify the developer that they should do
something more cross-platformish (to avoid the minor perf hit of the
wrapper).

(side note: I'm on vacation right now so don't have my normal test
setup; if things are broken then I'll fix when I get back).