WinJS.UI.ListView partial impl + listview sample
This commit includes a partial implementation of the WinJS.UI.ListView
class, and the Win8 SDK ListView template sample to test it.  The sample
is at parity with Win8 except for the need to tweak WinJS.UI.Page render
flow (currently flashes partial page).
Additional items in this commit:
* Created a shared "sdkCommonStyles.css" file that contains css tweaks
that address the lack of -ms-grid in the Win8 sdk samples.  Updated the
flipView sample to use it as well as the listView template sample.
* WinJS.UI.ListLayout class (to support WinJS.UI.ListView)
* Made the Rating.averageRating setter update the Rating control
automatically.
* Win8SDK's ListView Templates sample.  The only changes necessary to
make this work in bluesky are referencing the aforementioned sdk common
styles file, and swapping the WinJS references for bluesky.js
references.

Things *not* in this commit:
* ListView tests.  Those will come next.
* Grouped ListViews.