WinJS.UI.Flyout impl. + tests, plus DOMEventMixin
WinJS.UI.Flyout is implemented sufficiently for R1.  The only aspects
that are not at Win8 parity are minor:

* The flyout doesn't animate-in (because
WinJS.UI.Animation.show/hidePopup is NYI)
* The edge case where the flyout is explicitly shown to one side and
there isn't enough room on that side: Win8 resizes the flyout, while
bluesky just repositions it so that it fits on the screen.

This commit also includes:
* Tests for WinJS.UI.Flyout
* Complete DOMEventMixin implementation (but no tests yet)
* Made BaseControl derive from DOMEventMixin, and removed
no-longer-necessary event code from the various UI objects
* WinJS.Resources stub implementation