Event handling rewritten, plus flyout sample
Event handling is much closer to the win8 model now.  This checkin also
includes the Win8 SDK flyout control sample which is mostly working
except for one of the scenarios, and a bug in the flyout's light dismiss
handling (it dismisses the flyout even if you click within the flyout).