WinJS.UI.Fragments impl + tests (WIP)
This commit has a full WinJS.UI.Fragments implementation, tests for it,
and the Win8 SDK Fragment sample.

This is a WIP checkin with two bugs remaining to be fixed tomorrow:
* Loading a fragment a second time isn't quite working.  Reproable by
clicking the "load fragment" button twice in the fragment sample
scenario 1
* A bug in the Page rendering flow where Navigation.navigate is
returning before the DOM is fully set up.  repro'able by switching
scenarios in the fragment sample and watching the debug log.

Later, I'll refactor WinJS.UI.Pages to build on WinJS.UI.Fragments; for
now there's largely duplicate between the two.