Refactored WinJS.UI.Flyout's clickeater
Removed the old light-dismiss approach for WinJS.UI.Flyout and adopted
something closer to the WinJS.UI.AppBar model.  This fixes a number of
edge cases for left/right clicking when both and appbar and a flyout are
visible at the same time.

Also re-enabled the appbar in the stocksample app.