Windows.Storage done for R1!
Mental note: Windows.Storage is a LARGE NAMESPACE.  Wuff. Here's what
this commit includes:

* Windows.Storage namespace complete for R1.  This should work for the
majority of apps that work with text files.  There are still a plethora
of post-R1 things to do to this namespace, including:
- roaming files NYI (CachedFileManager is implemented, but haven't
hooked up backend services yet).
- no binary files/stream support
- custom folders (documents, images, etc) are only loosely supported.
* All Windows.Storage tests passing.  Up to 256 tests with this checkin;
how apropos!
* Windows.Storage.FileIO and PathIO implemented
* Windows.Storage.Search implemented, but only with defaultQuery (and
likely missing a swatch of deeper functionality).
* ApplicationData, ApplicationDataContainer,
ApplicationDataContainerSettings: these objects are implemented, but not
yet persisting.  Just need to figure out the right way to store them in
localStorage - but for now this should unblock any apps that use these
objects.
* SampleBrowser updated with fileAccess and applicationData samples

Looking forward to not thinking about Windows.Storage for a while!