Start of bluesky VS Addin
This commit includes the start of the bluesky "webify me" addin for
visual studio.  It is not yet ready for public use; there is no vsix and
it only works in a subset of cases.  But in those cases -  right click
on a WinJS project in VS' Solution Explorer, select "Win8 --> Bluesky",
and a moment later you have a web project ready to run on the web!