Make bluesky phoneGap-friendly
Made the following changes to enable Bluesky to integrate with PhoneGap
more smoothly:
* Added Bluesky.IsLocalExecution var that is set on init.  Set to true
for PhoneGap.  This impacts how paths are constructed
* Changing how paths are managed in WinJS.xhr, script loading, and more.
Still not handling upper/lower case cleanly throughout yet, so if you're
building for iOS (or other case-sensitive-file platform), you'll want to
just lower-case all filenames.
*

Note that phoneGap-Bluesky support is still in beta; it requires some
shenanigans on the app's part (especially on WP7).  Ping me
(jeffsim@bluesky.io) if you'd like to experiment with it and I'll write
it up.